♻️ refactor: move parametric dimension_of + conversions into unxts.parametric
