+++ packages/unxts.parametric/src/unxts/parametric/__init__.py
@@ -6,4 +6,8 @@ from ._src.base_parametric import AbstractParametricQuantity
 from ._src.parametric import PQ, ParametricQuantity
 
 # Import register modules for their dispatch/promotion side effects.
-from ._src import register_promotions  # noqa: F401
+from ._src import (  # noqa: F401
+    register_api,
+    register_conversions,
+    register_promotions,
+)
